Add hacs.json (#14)
* Initial commit.

* Scan all discovered neighbours.

Sanity check after scan is finished.

* Start initial scan after a delay.

* isort imports.

* Handle unknown neigbours.

* Handle failures gracefuly.

Don't include failed devices in next scan.
Don't scan devices not present in `zigbee.db`.

* scan_now service.

* Black format.

* Expose "current" topology through a property.

* Fix handling of devices not in db.

* Add WS command.

* "offline" property.

Neighbours are assumed to be online by default. Only neigbours which
were attempted to be scanned and failed are marked offline. This implies
that Zigbee End Devices are always online, because they are never
attempted to be scanned.

* Add "neighbours" missing in neighobour table.

Add devices as "offline" if present in zigbee.db, but not found in
neighbours.

* Make sure device_typs always has a value.

* Skip devices not supporting ZDO lqi_mgmt_req.

Samjin multisensor reports itself as a router, even though it is an end device.

* return timestamp in WS call (#2)

* ignore invalid neighbours reported by some devices (#1)

* Ignore 00:00:00:00:00:00:00:00 neighbours.

* Fix IEEE conversions.

* Use node descriptor only if available.

Derive device type from node descriptor (authorative) only if avaiable
ie device is online.

* Doh.

* Use hex representation for NWK addresses.

* Black format code.

* Update neighbour.py (#4)

prevent infinity loop

* fix unhadled exception on scan timeout (#5)

* Update README.md with usage instruction (#7)

Update README.md with usage instruction - Credits goes to eszkjc on Reddit

https://www.reddit.com/r/homeassistant/comments/eszkjc/zigbee_network_mapping_with_zha_my_77_node_network/

https://www.reddit.com/r/homeassistant/comments/eszkjc/zigbee_network_mapping_with_zha_my_77_node_network/

* Revert "fix unhadled exception on scan timeout (#5)" (#10)

This reverts commit 49fb8c76e7f96641effa9a2ab3a6d6e5376d650c.

* Create info.md

* Create LICENSE (#12)

* Update info.md

* Create hacs.json

Co-authored-by: Abílio Costa <abmantis@users.noreply.github.com>
Co-authored-by: jkkwon83 <jkkwon83@gmail.com>
Co-authored-by: Andreas Setterlind <Gamester17@users.noreply.github.com>